Connecting questions part.

+++ src/main/java/mk/edu/uklo/fikt/fiktexamweb/model/Question.java
@@ -14,19 +14,21 @@ public class Question {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private int id;
 	
 	private String text;
 	private String level;
 	private Long topicId;
-	
-	
-	public Long getId() {
+
+
+	public int getId() {
 		return id;
 	}
-	public void setId(Long id) {
+
+	public void setId(int id) {
 		this.id = id;
 	}
+
 	public String getText() {
 		return text;
 	}
